refactor: replace `utils/api` with `useFetchWithCsrf` composable across codebase
