+++ composables/countries.ts
@@ -1,4 +1,5 @@
 import type { ApiResponse } from '~/types';
+import { useFetchWithCsrf } from '~/composables/use-fetch-with-csrf';
 
 export interface Country {
   readonly code: string;
@@ -8,11 +9,11 @@ export interface Country {
 export function useCountries() {
   const countries = ref<Country[]>([]);
   const countriesLoadError = ref<string>('');
+  const { fetchWithCsrf } = useFetchWithCsrf();
 
   const loadCountries = async () => {
     try {
-      const response
-        = await fetchWithCsrf<ApiResponse<Country[]>>('/webapi/countries/');
+      const response = await fetchWithCsrf<ApiResponse<Country[]>>('/webapi/countries/');
       countries.value = response.result ?? [];
     }
     catch (error: any) {
