+++ store/payments/cards.ts
@@ -1,4 +1,5 @@
 import { set } from '@vueuse/core';
+import { useFetchWithCsrf } from '~/composables/use-fetch-with-csrf';
 import {
   type ApiResponse,
   type CreateCardNonceRequest,
@@ -7,13 +8,13 @@ import {
   SavedCard,
   SavedCardResponse,
 } from '~/types';
-import { fetchWithCsrf } from '~/utils/api';
 import { useLogger } from '~/utils/use-logger';
 
 export const usePaymentCardsStore = defineStore('payments/cards', () => {
   const card = ref<SavedCard>();
 
   const logger = useLogger('card-payment');
+  const { fetchWithCsrf } = useFetchWithCsrf();
 
   const getCard = async (): Promise<void> => {
     try {
